SpinnGymApplicationVertex
