+++ spinn_gym/games/__init__.py
@@ -12,3 +12,7 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+from .spinn_gym_application_vertex import SpinnGymApplicationVertex
+
+__all__ = ["SpinnGymApplicationVertex"]
