+++ spinn_gym/games/breakout/breakout.py
@@ -18,28 +18,13 @@ import numpy
 from spinn_utilities.overrides import overrides
 
 # PACMAN imports
-from pacman.model.constraints.key_allocator_constraints import (
-    ContiguousKeyRangeContraint)
-from pacman.model.graphs.application.abstract import (
-    AbstractOneAppOneMachineVertex)
 from pacman.model.graphs.common import Slice
-from spinn_utilities.config_holder import get_config_int
-
-# SpinnFrontEndCommon imports
-from spinn_front_end_common.abstract_models import AbstractChangableAfterRun
-from spinn_front_end_common.abstract_models.abstract_has_associated_binary \
-    import AbstractHasAssociatedBinary
-from spinn_front_end_common.abstract_models. \
-    abstract_provides_outgoing_partition_constraints import \
-    AbstractProvidesOutgoingPartitionConstraints
-from spinn_front_end_common.utilities.utility_objs import ExecutableType
 
 # sPyNNaker imports
-from spynnaker.pyNN.models.abstract_models import \
-    AbstractAcceptsIncomingSynapses
 from spynnaker.pyNN.models.common import AbstractNeuronRecordable
-from spynnaker.pyNN.models.common.simple_population_settable \
-    import SimplePopulationSettable
+
+# common imports
+from spinn_gym.games import SpinnGymApplicationVertex
 
 # Breakout imports
 from spinn_gym.games.breakout.breakout_machine_vertex import \
@@ -49,34 +34,7 @@ from spinn_gym.games.breakout.breakout_machine_vertex import \
 # ----------------------------------------------------------------------------
 # Breakout
 # ----------------------------------------------------------------------------
-class Breakout(AbstractOneAppOneMachineVertex,
-               AbstractProvidesOutgoingPartitionConstraints,
-               AbstractAcceptsIncomingSynapses,
-               AbstractNeuronRecordable,
-               SimplePopulationSettable):
-
-    @overrides(AbstractAcceptsIncomingSynapses.verify_splitter)
-    def verify_splitter(self, splitter):
-        # Need to ignore this verify
-        pass
-
-    @overrides(AbstractAcceptsIncomingSynapses.get_connections_from_machine)
-    def get_connections_from_machine(
-            self, transceiver, placements, app_edge, synapse_info):
-
-        # TODO: make this work properly (the following call does nothing)
-
-        super(Breakout, self).get_connections_from_machine(
-            transceiver, placements, app_edge, synapse_info)
-
-    def set_synapse_dynamics(self, synapse_dynamics):
-        pass
-
-    def clear_connection_cache(self):
-        pass
-
-    def get_synapse_id_by_target(self, target):
-        return 0
+class Breakout(SpinnGymApplicationVertex):
 
     BREAKOUT_REGION_BYTES = 4
     PARAM_REGION_BYTES = 40
@@ -156,88 +114,10 @@ class Breakout(AbstractOneAppOneMachineVertex,
                 incoming_spike_buffer_size, simulation_duration_ms, bricking,
                 random_seed),
             label=label, constraints=constraints)
-        AbstractProvidesOutgoingPartitionConstraints.__init__(self)
-        SimplePopulationSettable.__init__(self)
-        AbstractChangableAfterRun.__init__(self)
-        AbstractAcceptsIncomingSynapses.__init__(self)
-        self._change_requires_mapping = True
-        if incoming_spike_buffer_size is None:
-            self._incoming_spike_buffer_size = get_config_int(
-                "Simulation", "incoming_spike_buffer_size")
-
-    def neurons(self):
-        return self._n_neurons
-
-    @property
-    @overrides(AbstractOneAppOneMachineVertex.n_atoms)
-    def n_atoms(self):
-
-        # **TODO** should we calculate this automatically
-        # based on log2 of width and height?
-        return self._n_neurons
-
-    # ------------------------------------------------------------------------
-    # AbstractHasAssociatedBinary overrides
-    # ------------------------------------------------------------------------
-    @overrides(AbstractHasAssociatedBinary.get_binary_file_name)
-    def get_binary_file_name(self):
-        print("Breakout get_binary_file_name")
-        return "breakout.aplx"
-
-    @overrides(AbstractHasAssociatedBinary.get_binary_start_type)
-    def get_binary_start_type(self):
-        print("Breakout get_binary_start_type")
-        # return ExecutableStartType.USES_SIMULATION_INTERFACE
-        return ExecutableType.USES_SIMULATION_INTERFACE
-
-    # ------------------------------------------------------------------------
-    # AbstractProvidesOutgoingPartitionConstraints overrides
-    # ------------------------------------------------------------------------
-    @overrides(AbstractProvidesOutgoingPartitionConstraints.
-               get_outgoing_partition_constraints)
-    def get_outgoing_partition_constraints(self, partition):
-        return [ContiguousKeyRangeContraint()]
-
-    @property
-    @overrides(AbstractChangableAfterRun.requires_mapping)
-    def requires_mapping(self):
-        return self._change_requires_mapping
-
-    @overrides(AbstractChangableAfterRun.mark_no_changes)
-    def mark_no_changes(self):
-        self._change_requires_mapping = False
-
-    @overrides(SimplePopulationSettable.set_value)
-    def set_value(self, key, value):
-        SimplePopulationSettable.set_value(self, key, value)
-        self._change_requires_neuron_parameters_reload = True
 
     # ------------------------------------------------------------------------
     # Recording overrides
     # ------------------------------------------------------------------------
-    @overrides(
-        AbstractNeuronRecordable.clear_recording)
-    def clear_recording(
-            self, variable, buffer_manager, placements):
-        self._clear_recording_region(buffer_manager, placements, 0)
-
-    @overrides(AbstractNeuronRecordable.get_recordable_variables)
-    def get_recordable_variables(self):
-        return 'score'
-
-    @overrides(AbstractNeuronRecordable.is_recording)
-    def is_recording(self, variable):
-        return True
-
-    @overrides(AbstractNeuronRecordable.set_recording)
-    def set_recording(self, variable, new_state=True, sampling_interval=None,
-                      indexes=None):
-        pass
-
-    @overrides(AbstractNeuronRecordable.get_neuron_sampling_interval)
-    def get_neuron_sampling_interval(self, variable):
-        return 10000  # 10 seconds hard coded in bkout.c
-
     @overrides(AbstractNeuronRecordable.get_data)
     def get_data(
             self, variable, n_machine_time_steps, placements, buffer_manager):
@@ -254,26 +134,3 @@ class Breakout(AbstractOneAppOneMachineVertex,
         output_data = numpy.array(data, dtype=numpy.uint8).view(numpy_format)
 
         return output_data
-
-    def _clear_recording_region(
-            self, buffer_manager, placements, recording_region_id):
-        """ Clear a recorded data region from the buffer manager.
-
-        :param buffer_manager: the buffer manager object
-        :param placements: the placements object
-        :param recording_region_id: the recorded region ID for clearing
-        :rtype: None
-        """
-        for machine_vertex in self.machine_vertices:
-            placement = placements.get_placement_of_vertex(machine_vertex)
-            buffer_manager.clear_recorded_data(
-                placement.x, placement.y, placement.p, recording_region_id)
-
-    def reset_ring_buffer_shifts(self):
-        pass
-
-    def __str__(self):
-        return "{} with {} atoms".format(self._label, self.n_atoms)
-
-    def __repr__(self):
-        return self.__str__()
